migrate to pyproject.toml (PEP 621)
