+++ MANIFEST.in
@@ -1,7 +1,6 @@
 graft artiq/firmware
 graft artiq/examples
 include artiq/gui/logo*.svg
-include versioneer.py
 include artiq/_version.py
 include artiq/coredevice/coredevice_generic.schema.json
 include artiq/compiler/kernel.ld
